+++ pages/classroom/[classroomId]/study/[assignmentId].tsx
@@ -92,7 +92,7 @@ export const getServerSideProps: GetServerSideProps = async (context) => {
 
   const course = await prisma.course.findUnique({
     where: { id: courseId as string },
-    select: { title: true },
+    select: { title: true, privacySetting: true },
   });
 
   return {
@@ -101,6 +101,7 @@ export const getServerSideProps: GetServerSideProps = async (context) => {
         reviewInstances,
         title: course?.title,
         intervals,
+        isSharedCourse: course?.privacySetting !== "PRIVATE",
       })
     ) as StudyFlashcardsPageProps,
   };